Refactor onboarding sequence into modular components
