+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -169,6 +169,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 722 to 55 lines by extracting manager, AI processor, coordinator, and config modules. New orchestrator coordinates components and maintains functionality.
 
+### MODERATE-016: V2 Onboarding Sequence Modularization ✅
+- **File**: `src/core/v2_onboarding_sequence.py`
+- **Status**: Completed
+- **Assigned To**: Agent-5
+- **Completion Date**: 2025-08-24
+- **Summary**: Split monolithic onboarding sequence into core, validator, coordinator, and config modules with a lightweight orchestrator. Imports updated and tests verified.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
