+++ src/core/v2_onboarding_sequence.py
@@ -1,661 +1,35 @@
-#!/usr/bin/env python3
-"""
-V2 Onboarding Sequence - Agent Cellphone V2
-===========================================
+"""Orchestrator for the V2 Onboarding Sequence.
 
-Advanced onboarding sequence for real agent communication system V2.
-Follows V2 standards: ≤300 LOC, OOP design, SRP.
-
-Author: V2 Onboarding & Communication Specialist
-License: MIT
+This module wires together the modular components responsible for
+onboarding new agents into the V2 system.
 """
 
-import logging
-import threading
-import time
-import json
-import uuid
-
-from src.utils.stability_improvements import stability_manager, safe_import
-from datetime import datetime, timedelta
-from typing import Dict, Any, List, Optional, Callable
-from dataclasses import dataclass, field
-from enum import Enum
-from pathlib import Path
-
-# Import V2 components
-from .communication_compatibility_layer import (
-    AgentCommunicationProtocol,
-    MessageType,
-    MessagePriority,
-    InboxManager
+from .v2_onboarding_sequence_core import (
+    V2OnboardingSequenceCore,
+    OnboardingStatus,
+    OnboardingPhase,
+    OnboardingSession,
+    OnboardingMessage,
 )
-from .fsm_core_v2 import FSMCoreV2
-
-# Configure logging
-logger = logging.getLogger(__name__)
-
-
-class OnboardingStatus(Enum):
-    """Onboarding status values"""
-
-    PENDING = "pending"
-    INITIALIZING = "initializing"
-    TRAINING = "training"
-    VALIDATION = "validation"
-    COMPLETED = "completed"
-    FAILED = "failed"
-
-
-class OnboardingPhase(Enum):
-    """Onboarding phases"""
-
-    SYSTEM_OVERVIEW = "system_overview"
-    ROLE_ASSIGNMENT = "role_assignment"
-    CAPABILITY_TRAINING = "capability_training"
-    INTEGRATION_TESTING = "integration_testing"
-    PERFORMANCE_VALIDATION = "performance_validation"
-    READINESS_CONFIRMATION = "readiness_confirmation"
-
-
-@dataclass
-class OnboardingSession:
-    """Onboarding session tracking"""
-
-    session_id: str
-    agent_id: str
-    status: OnboardingStatus
-    current_phase: OnboardingPhase
-    completed_phases: List[OnboardingPhase]
-    start_time: datetime
-    completion_time: Optional[datetime] = None
-    performance_metrics: Dict[str, Any] = field(default_factory=dict)
-    validation_results: Dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class OnboardingMessage:
-    """Structured onboarding message"""
-
-    message_id: str
-    phase: OnboardingPhase
-    content: str
-    role_specific: bool
-    requires_response: bool
-    validation_criteria: Dict[str, Any] = field(default_factory=dict)
-
-
-class V2OnboardingSequence:
-    """
-    V2 Onboarding Sequence - Single responsibility: Agent onboarding and training.
-
-    Follows V2 standards: ≤300 LOC, OOP design, SRP.
-    """
-
-    def __init__(self, config: Dict[str, Any] = None):
-        """Initialize the V2 onboarding sequence"""
-        self.config = config or {}
-
-        # Core components
-        self.communication_protocol: Optional[AgentCommunicationProtocol] = None
-        self.fsm_core: Optional[FSMCoreV2] = None
-        self.inbox_manager: Optional[InboxManager] = None
-
-        # Onboarding state
-        self.active_sessions: Dict[str, OnboardingSession] = {}
-        self.onboarding_templates: Dict[str, OnboardingMessage] = {}
-        self.role_definitions: Dict[str, Dict[str, Any]] = {}
-
-        # Configuration
-        self.phase_timeout = self.config.get("phase_timeout", 300)  # 5 minutes
-        self.validation_retries = self.config.get("validation_retries", 3)
-        self.performance_thresholds = self.config.get("performance_thresholds", {})
-
-        # Initialize onboarding system
-        self._initialize_onboarding_system()
-
-        logger.info("V2OnboardingSequence initialized")
-
-    def _initialize_onboarding_system(self):
-        """Initialize the onboarding system components"""
-        try:
-            # Initialize role definitions for V2 system
-            self.role_definitions = {
-                "Agent-1": {
-                    "role": "System Coordinator & Project Manager",
-                    "capabilities": [
-                        "FSM coordination",
-                        "Multi-repository oversight",
-                        "Performance monitoring",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-                "Agent-2": {
-                    "role": "Frontend Development Specialist",
-                    "capabilities": [
-                        "Modern web development",
-                        "UI/UX design",
-                        "Performance optimization",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-                "Agent-3": {
-                    "role": "Backend Development Specialist",
-                    "capabilities": [
-                        "API development",
-                        "Database design",
-                        "Security implementation",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-                "Agent-4": {
-                    "role": "DevOps & Infrastructure Specialist",
-                    "capabilities": [
-                        "Cloud infrastructure",
-                        "CI/CD pipelines",
-                        "Monitoring systems",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-                "Agent-5": {
-                    "role": "Gaming & Entertainment Specialist",
-                    "capabilities": [
-                        "Game development",
-                        "Entertainment systems",
-                        "Performance optimization",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-                "Agent-6": {
-                    "role": "AI/ML & Research Specialist",
-                    "capabilities": [
-                        "Machine learning",
-                        "AI system integration",
-                        "Research initiatives",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-                "Agent-7": {
-                    "role": "Web & UI Framework Specialist",
-                    "capabilities": [
-                        "Framework development",
-                        "Component systems",
-                        "Performance optimization",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-                "Agent-8": {
-                    "role": "Mobile & Cross-Platform Specialist",
-                    "capabilities": [
-                        "Mobile development",
-                        "Cross-platform compatibility",
-                        "Performance optimization",
-                    ],
-                    "onboarding_phases": [
-                        OnboardingPhase.SYSTEM_OVERVIEW,
-                        OnboardingPhase.ROLE_ASSIGNMENT,
-                        OnboardingPhase.CAPABILITY_TRAINING,
-                        OnboardingPhase.INTEGRATION_TESTING,
-                    ],
-                },
-            }
-
-            # Initialize onboarding message templates
-            self._initialize_message_templates()
-
-            logger.info("Onboarding system components initialized")
-
-        except Exception as e:
-            logger.error(f"Failed to initialize onboarding system: {e}")
-            raise
-
-    def _initialize_message_templates(self):
-        """Initialize onboarding message templates"""
-        try:
-            # System overview template
-            self.onboarding_templates["system_overview"] = OnboardingMessage(
-                message_id=str(uuid.uuid4()),
-                phase=OnboardingPhase.SYSTEM_OVERVIEW,
-                content="🚀 Welcome to Agent Cellphone V2! You are now part of the most advanced autonomous agent coordination platform ever created. The V2 system features advanced FSM-driven workflows, real-time coordination, and enterprise-grade capabilities.",
-                role_specific=False,
-                requires_response=True,
-                validation_criteria={"response_time": 30, "comprehension": True},
-            )
-
-            # Role assignment template
-            self.onboarding_templates["role_assignment"] = OnboardingMessage(
-                message_id=str(uuid.uuid4()),
-                phase=OnboardingPhase.ROLE_ASSIGNMENT,
-                content="🎯 Role Assignment: You have been assigned a specialized role in the V2 system. This role leverages your unique capabilities and integrates with the advanced FSM coordination engine.",
-                role_specific=True,
-                requires_response=True,
-                validation_criteria={
-                    "role_acceptance": True,
-                    "capability_confirmation": True,
-                },
-            )
-
-            # Capability training template
-            self.onboarding_templates["capability_training"] = OnboardingMessage(
-                message_id=str(uuid.uuid4()),
-                phase=OnboardingPhase.CAPABILITY_TRAINING,
-                content="🔧 Capability Training: You will now receive specialized training for your role, including V2 development standards, FSM integration, and performance optimization techniques.",
-                role_specific=True,
-                requires_response=True,
-                validation_criteria={
-                    "training_completion": True,
-                    "skill_assessment": True,
-                },
-            )
-
-            # Integration testing template
-            self.onboarding_templates["integration_testing"] = OnboardingMessage(
-                message_id=str(uuid.uuid4()),
-                phase=OnboardingPhase.INTEGRATION_TESTING,
-                content="🧪 Integration Testing: You will now participate in integration testing to validate your coordination with other agents and the FSM system.",
-                role_specific=False,
-                requires_response=True,
-                validation_criteria={
-                    "integration_success": True,
-                    "coordination_effectiveness": True,
-                },
-            )
-
-            logger.info("Onboarding message templates initialized")
-
-        except Exception as e:
-            logger.error(f"Failed to initialize message templates: {e}")
-            raise
-
-    def start_onboarding(
-        self,
-        agent_id: str,
-        communication_protocol: AgentCommunicationProtocol,
-        fsm_core: FSMCoreV2,
-        inbox_manager: InboxManager,
-    ) -> str:
-        """Start the onboarding process for an agent"""
-        try:
-            # Validate agent ID
-            if agent_id not in self.role_definitions:
-                raise ValueError(f"Unknown agent ID: {agent_id}")
-
-            # Store component references
-            self.communication_protocol = communication_protocol
-            self.fsm_core = fsm_core
-            self.inbox_manager = inbox_manager
-
-            # Create onboarding session
-            session_id = f"onboard_{agent_id}_{int(time.time())}"
-            session = OnboardingSession(
-                session_id=session_id,
-                agent_id=agent_id,
-                status=OnboardingStatus.INITIALIZING,
-                current_phase=OnboardingPhase.SYSTEM_OVERVIEW,
-                completed_phases=[],
-                start_time=datetime.now(),
-            )
-
-            self.active_sessions[session_id] = session
-
-            # Start onboarding in background thread
-            onboarding_thread = threading.Thread(
-                target=self._execute_onboarding_sequence,
-                args=(session_id,),
-                daemon=True,
-            )
-            onboarding_thread.start()
-
-            logger.info(f"Started onboarding session {session_id} for agent {agent_id}")
-            return session_id
-
-        except Exception as e:
-            logger.error(f"Failed to start onboarding for {agent_id}: {e}")
-            return ""
-
-    def _execute_onboarding_sequence(self, session_id: str):
-        """Execute the complete onboarding sequence for an agent"""
-        try:
-            session = self.active_sessions[session_id]
-            agent_id = session.agent_id
-            role_info = self.role_definitions[agent_id]
-
-            logger.info(f"Executing onboarding sequence for {agent_id}")
-
-            # Phase 1: System Overview
-            if self._execute_phase(session, OnboardingPhase.SYSTEM_OVERVIEW):
-                session.completed_phases.append(OnboardingPhase.SYSTEM_OVERVIEW)
-
-                # Phase 2: Role Assignment
-                if self._execute_phase(session, OnboardingPhase.ROLE_ASSIGNMENT):
-                    session.completed_phases.append(OnboardingPhase.ROLE_ASSIGNMENT)
-
-                    # Phase 3: Capability Training
-                    if self._execute_phase(
-                        session, OnboardingPhase.CAPABILITY_TRAINING
-                    ):
-                        session.completed_phases.append(
-                            OnboardingPhase.CAPABILITY_TRAINING
-                        )
-
-                        # Phase 4: Integration Testing
-                        if self._execute_phase(
-                            session, OnboardingPhase.INTEGRATION_TESTING
-                        ):
-                            session.completed_phases.append(
-                                OnboardingPhase.INTEGRATION_TESTING
-                            )
-
-                            # Final validation and completion
-                            if self._validate_onboarding_completion(session):
-                                session.status = OnboardingStatus.COMPLETED
-                                session.completion_time = datetime.now()
-                                logger.info(
-                                    f"Onboarding completed successfully for {agent_id}"
-                                )
-                            else:
-                                session.status = OnboardingStatus.FAILED
-                                logger.error(
-                                    f"Onboarding validation failed for {agent_id}"
-                                )
-                        else:
-                            session.status = OnboardingStatus.FAILED
-                            logger.error(f"Integration testing failed for {agent_id}")
-                    else:
-                        session.status = OnboardingStatus.FAILED
-                        logger.error(f"Capability training failed for {agent_id}")
-                else:
-                    session.status = OnboardingStatus.FAILED
-                    logger.error(f"Role assignment failed for {agent_id}")
-            else:
-                session.status = OnboardingStatus.FAILED
-                logger.error(f"System overview failed for {agent_id}")
-
-        except Exception as e:
-            logger.error(
-                f"Onboarding sequence execution failed for session {session_id}: {e}"
-            )
-            if session_id in self.active_sessions:
-                self.active_sessions[session_id].status = OnboardingStatus.FAILED
-
-    def _execute_phase(
-        self, session: OnboardingSession, phase: OnboardingPhase
-    ) -> bool:
-        """Execute a specific onboarding phase"""
-        try:
-            session.current_phase = phase
-            agent_id = session.agent_id
-
-            logger.info(f"Executing phase {phase.value} for {agent_id}")
-
-            # Get phase-specific message
-            message = self._get_phase_message(phase, agent_id)
-
-            # Send message through communication protocol
-            if self.communication_protocol:
-                message_id = self.communication_protocol.send_message(
-                    sender_id="SYSTEM",
-                    recipient_id=agent_id,
-                    message_type=MessageType.COORDINATION,
-                    payload={
-                        "phase": phase.value,
-                        "content": message.content,
-                        "requires_response": message.requires_response,
-                        "validation_criteria": message.validation_criteria,
-                    },
-                    priority=MessagePriority.HIGH,
-                )
-
-                # Wait for response and validate
-                if message.requires_response:
-                    return self._wait_for_phase_response(session, phase, message_id)
-                else:
-                    return True
-
-            return False
-
-        except Exception as e:
-            logger.error(f"Phase execution failed for {phase.value}: {e}")
-            return False
-
-    def _get_phase_message(
-        self, phase: OnboardingPhase, agent_id: str
-    ) -> OnboardingMessage:
-        """Get the message for a specific phase"""
-        try:
-            if phase == OnboardingPhase.SYSTEM_OVERVIEW:
-                return self.onboarding_templates["system_overview"]
-            elif phase == OnboardingPhase.ROLE_ASSIGNMENT:
-                role_info = self.role_definitions[agent_id]
-                message = OnboardingMessage(
-                    message_id=str(uuid.uuid4()),
-                    phase=phase,
-                    content=f"🎯 Role Assignment: You are {role_info['role']}. Your capabilities include: {', '.join(role_info['capabilities'])}. Are you ready to accept this role?",
-                    role_specific=True,
-                    requires_response=True,
-                    validation_criteria={
-                        "role_acceptance": True,
-                        "capability_confirmation": True,
-                    },
-                )
-                return message
-            elif phase == OnboardingPhase.CAPABILITY_TRAINING:
-                role_info = self.role_definitions[agent_id]
-                message = OnboardingMessage(
-                    message_id=str(uuid.uuid4()),
-                    phase=phase,
-                    content=f"🔧 Capability Training: You will now receive specialized training for your role as {role_info['role']}. This includes V2 development standards, FSM integration, and performance optimization.",
-                    role_specific=True,
-                    requires_response=True,
-                    validation_criteria={
-                        "training_completion": True,
-                        "skill_assessment": True,
-                    },
-                )
-                return message
-            elif phase == OnboardingPhase.INTEGRATION_TESTING:
-                message = OnboardingMessage(
-                    message_id=str(uuid.uuid4()),
-                    phase=phase,
-                    content="🧪 Integration Testing: You will now participate in integration testing to validate your coordination with other agents and the FSM system. Ready to begin?",
-                    role_specific=False,
-                    requires_response=True,
-                    validation_criteria={
-                        "integration_success": True,
-                        "coordination_effectiveness": True,
-                    },
-                )
-                return message
-            else:
-                # Default message
-                return self.onboarding_templates["system_overview"]
-
-        except Exception as e:
-            logger.error(f"Failed to get phase message: {e}")
-            return self.onboarding_templates["system_overview"]
-
-    def _wait_for_phase_response(
-        self, session: OnboardingSession, phase: OnboardingPhase, message_id: str
-    ) -> bool:
-        """Wait for and validate phase response"""
-        try:
-            start_time = time.time()
-            timeout = self.phase_timeout
-
-            while time.time() - start_time < timeout:
-                # Check for response in inbox
-                if self.inbox_manager:
-                    # This would check for responses from the agent
-                    # For now, simulate successful response
-                    time.sleep(1)
-                    return True
-
-                time.sleep(0.1)
-
-            logger.warning(f"Phase {phase.value} timed out for {session.agent_id}")
-            return False
-
-        except Exception as e:
-            logger.error(f"Phase response validation failed: {e}")
-            return False
-
-    def _validate_onboarding_completion(self, session: OnboardingSession) -> bool:
-        """Validate that onboarding has been completed successfully"""
-        try:
-            # Check if all required phases are completed
-            required_phases = self.role_definitions[session.agent_id][
-                "onboarding_phases"
-            ]
-            all_phases_completed = all(
-                phase in session.completed_phases for phase in required_phases
-            )
-
-            if not all_phases_completed:
-                logger.warning(
-                    f"Not all required phases completed for {session.agent_id}"
-                )
-                return False
-
-            # Validate performance metrics
-            if not self._validate_performance_metrics(session):
-                logger.warning(f"Performance validation failed for {session.agent_id}")
-                return False
-
-            # Create FSM task for the agent
-            if self.fsm_core:
-                task_id = self.fsm_core.create_task(
-                    title=f"Onboarding Completion - {session.agent_id}",
-                    description=f"Agent {session.agent_id} has completed onboarding and is ready for active participation",
-                    assigned_agent=session.agent_id,
-                )
-                logger.info(f"Created FSM task {task_id} for onboarding completion")
-
-            return True
-
-        except Exception as e:
-            logger.error(f"Onboarding completion validation failed: {e}")
-            return False
-
-    def _validate_performance_metrics(self, session: OnboardingSession) -> bool:
-        """Validate performance metrics for onboarding completion"""
-        try:
-            # For now, return True as basic validation
-            # In a real implementation, this would check actual performance metrics
-            return True
-
-        except Exception as e:
-            logger.error(f"Performance validation failed: {e}")
-            return False
-
-    def get_onboarding_status(self, session_id: str) -> Optional[Dict[str, Any]]:
-        """Get the status of an onboarding session"""
-        try:
-            if session_id not in self.active_sessions:
-                return None
-
-            session = self.active_sessions[session_id]
-            return {
-                "session_id": session.session_id,
-                "agent_id": session.agent_id,
-                "status": session.status.value,
-                "current_phase": session.current_phase.value,
-                "completed_phases": [phase.value for phase in session.completed_phases],
-                "start_time": session.start_time.isoformat(),
-                "completion_time": session.completion_time.isoformat()
-                if session.completion_time
-                else None,
-                "performance_metrics": session.performance_metrics,
-                "validation_results": session.validation_results,
-            }
-
-        except Exception as e:
-            logger.error(f"Failed to get onboarding status: {e}")
-            return None
-
-    def get_all_onboarding_status(self) -> Dict[str, Any]:
-        """Get status of all onboarding sessions"""
-        try:
-            return {
-                session_id: self.get_onboarding_status(session_id)
-                for session_id in self.active_sessions.keys()
-            }
-
-        except Exception as e:
-            logger.error(f"Failed to get all onboarding status: {e}")
-            return {}
-
-    def cleanup_completed_sessions(self):
-        """Clean up completed onboarding sessions"""
-        try:
-            completed_sessions = [
-                session_id
-                for session_id, session in self.active_sessions.items()
-                if session.status
-                in [OnboardingStatus.COMPLETED, OnboardingStatus.FAILED]
-            ]
-
-            for session_id in completed_sessions:
-                del self.active_sessions[session_id]
-
-            if completed_sessions:
-                logger.info(f"Cleaned up {len(completed_sessions)} completed sessions")
-
-        except Exception as e:
-            logger.error(f"Failed to cleanup completed sessions: {e}")
-
-
-# Example usage and testing
-if __name__ == "__main__":
-    # Initialize onboarding sequence
-    onboarding = V2OnboardingSequence()
-
-    # Start onboarding for an agent
-    session_id = onboarding.start_onboarding("Agent-1", None, None, None)
-
-    if session_id:
-        print(f"Onboarding started with session ID: {session_id}")
-
-        # Monitor progress
-        while True:
-            status = onboarding.get_onboarding_status(session_id)
-            if status and status["status"] in ["completed", "failed"]:
-                print(f"Onboarding {status['status']} for {status['agent_id']}")
-                break
-            time.sleep(1)
-    else:
-        print("Failed to start onboarding")
+from .v2_onboarding_sequence_validator import V2OnboardingSequenceValidator
+from .v2_onboarding_sequence_coordinator import V2OnboardingSequenceCoordinator
+
+
+class V2OnboardingSequence(
+    V2OnboardingSequenceValidator,
+    V2OnboardingSequenceCoordinator,
+    V2OnboardingSequenceCore,
+):
+    """Main entry point combining all onboarding components."""
+
+    def __init__(self, config=None):
+        V2OnboardingSequenceCore.__init__(self, config)
+
+
+__all__ = [
+    "V2OnboardingSequence",
+    "OnboardingStatus",
+    "OnboardingPhase",
+    "OnboardingSession",
+    "OnboardingMessage",
+]
